Client works well, but needs more work with the setup.

+++ src/main/java/se/helgestenstrom/ResourceRecord.java
@@ -7,7 +7,7 @@ public class ResourceRecord {
     private final Name name;
     private final ByteList rData;
     private final int rDataClass;
-    private final int timeToLive;
+    private final long timeToLive;
     private final int type;
 
 
@@ -18,7 +18,7 @@ public class ResourceRecord {
      * @param timeToLive u16 of the TTL, Time to Live, according to RFC 1035, section 4.1.3
      * @param rData the RDATA according to RFC 1035, section 4.1.3
      */
-    public ResourceRecord(Name name, int type, int rDataClass, int timeToLive, ByteList rData) {
+    public ResourceRecord(Name name, int type, int rDataClass, long timeToLive, ByteList rData) {
         this.name = name;
         this.type = type;
         this.rDataClass = rDataClass;
@@ -38,7 +38,7 @@ public class ResourceRecord {
         return rDataClass;
     }
 
-    public int getTimeToLive() {
+    public long getTimeToLive() {
         return timeToLive;
     }
